additional_requirements.txt
# Complete requirements for ML Trading System
# Core ML and Data Science
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
numba>=0.57.0

# Technical Analysis
talib-binary>=0.4.25

# Web Framework and ASGI
Flask>=2.3.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
gunicorn>=21.2.0
gevent>=23.7.0
hypercorn>=0.14.4
asgiref>=3.7.0

# Task Queue
celery>=5.3.0
redis>=4.6.0
kombu>=5.3.0

# Hyperparameter Optimization
optuna>=3.4.0

# Data Sources and HTTP
yfinance>=0.2.20
requests>=2.31.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Google Cloud Storage
google-cloud-storage>=2.10.0
google-auth>=2.23.0
google-api-core>=2.11.0

# Data Serialization
joblib>=1.3.0
pickle5>=0.0.12

# Monitoring and Metrics
prometheus-client>=0.17.0
prometheus-flask-exporter>=0.23.0
psutil>=5.9.0

# Security
cryptography>=41.0.0
marshmallow>=3.20.0
werkzeug>=2.3.0
bcrypt>=4.0.0

# Utilities
python-dotenv>=1.0.0
networkx>=3.1.0
click>=8.1.0
python-dateutil>=2.8.0
pytz>=2023.3

# Financial Analysis
backtrader>=1.9.78
empyrical>=0.5.5

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter for development
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional async support
asyncpg>=0.28.0  # For async database if needed
aiofiles>=23.0.0  # For async file operations

# Market data alternatives
alpha-vantage>=2.3.1
iexfinance>=0.5.0
quandl>=3.7.0

# Memory optimization
memory-profiler>=0.61.0
pympler>=0.9

# Production deployment
supervisor>=4.2.5  # Process management
nginx>=1.0.0  # Reverse proxy (system package)

# Optional: GPU support
# torch-audio>=2.0.0  # Uncomment if needed
# torch-vision>=0.15.0  # Uncomment if needed
